+++ crates/ika-move-packages/packages/ika_system/sources/system.move
@@ -456,7 +456,7 @@ public fun request_dwallet_network_decryption_key_dkg_by_cap(
 public fun set_supported_curves_to_signature_algorithms_to_hash_schemes(
     self: &mut System,
     dwallet_2pc_mpc_coordinator: &mut DWalletCoordinator,
-    supported_curves_to_signature_algorithms_to_hash_schemes: VecMap<u8, VecMap<u8, vector<u8>>>,
+    supported_curves_to_signature_algorithms_to_hash_schemes: VecMap<u32, VecMap<u32, vector<u32>>>,
     protocol_cap: &ProtocolCap,
 ) {
     let dwallet_2pc_mpc_coordinator_inner = dwallet_2pc_mpc_coordinator.inner_mut();
@@ -466,9 +466,9 @@ public fun set_supported_curves_to_signature_algorithms_to_hash_schemes(
 public fun set_paused_curves_and_signature_algorithms(
     self: &mut System,
     dwallet_2pc_mpc_coordinator: &mut DWalletCoordinator,
-    paused_curves: vector<u8>,
-    paused_signature_algorithms: vector<u8>,
-    paused_hash_schemes: vector<u8>,
+    paused_curves: vector<u32>,
+    paused_signature_algorithms: vector<u32>,
+    paused_hash_schemes: vector<u32>,
     protocol_cap: &ProtocolCap,
 ) {
     let dwallet_2pc_mpc_coordinator_inner = dwallet_2pc_mpc_coordinator.inner_mut();
